+++ docs/js/metro.js
@@ -3155,13 +3155,13 @@ $.widget("metro.calendar", {
             table = this.element.find('.calendar-grid');
 
         if (this._mode === 'day') {
-            table.find('.btn-select-month').on('click', function (e) {
+            table.find('.btn-select-month').off('click').on('click', function (e) {
                 e.preventDefault();
                 e.stopPropagation();
                 that._mode = 'month';
                 that._renderCalendar();
             });
-            table.find('.btn-previous-month').on('click', function (e) {
+            table.find('.btn-previous-month').off('click').on('click', function (e) {
                 that._event = 'eventPrevious';
                 e.preventDefault();
                 e.stopPropagation();
@@ -3172,7 +3172,7 @@ $.widget("metro.calendar", {
                 }
                 that._renderCalendar();
             });
-            table.find('.btn-next-month').on('click', function (e) {
+            table.find('.btn-next-month').off('click').on('click', function (e) {
                 that._event = 'eventNext';
                 e.preventDefault();
                 e.stopPropagation();
@@ -3183,14 +3183,14 @@ $.widget("metro.calendar", {
                 }
                 that._renderCalendar();
             });
-            table.find('.btn-previous-year').on('click', function (e) {
+            table.find('.btn-previous-year').off('click').on('click', function (e) {
                 that._event = 'eventPrevious';
                 e.preventDefault();
                 e.stopPropagation();
                 that._year -= 1;
                 that._renderCalendar();
             });
-            table.find('.btn-next-year').on('click', function (e) {
+            table.find('.btn-next-year').off('click').on('click', function (e) {
                 that._event = 'eventNext';
                 e.preventDefault();
                 e.stopPropagation();
@@ -3236,7 +3236,7 @@ $.widget("metro.calendar", {
                 }
             });
         } else if (this._mode === 'month') {
-            table.find('.month a').on('click', function (e) {
+            table.find('.month a').off('click').on('click', function (e) {
                 that._event = 'eventNext';
                 e.preventDefault();
                 e.stopPropagation();
@@ -3244,21 +3244,21 @@ $.widget("metro.calendar", {
                 that._mode = 'day';
                 that._renderCalendar();
             });
-            table.find('.btn-previous-year').on('click', function (e) {
+            table.find('.btn-previous-year').off('click').on('click', function (e) {
                 that._event = 'eventPrevious';
                 e.preventDefault();
                 e.stopPropagation();
                 that._year -= 1;
                 that._renderCalendar();
             });
-            table.find('.btn-next-year').on('click', function (e) {
+            table.find('.btn-next-year').off('click').on('click', function (e) {
                 that._event = 'eventNext';
                 e.preventDefault();
                 e.stopPropagation();
                 that._year += 1;
                 that._renderCalendar();
             });
-            table.find('.btn-select-year').on('click', function (e) {
+            table.find('.btn-select-year').off('click').on('click', function (e) {
                 that._event = 'eventNext';
                 e.preventDefault();
                 e.stopPropagation();
@@ -3266,7 +3266,7 @@ $.widget("metro.calendar", {
                 that._renderCalendar();
             });
         } else {
-            table.find('.year a').on('click', function (e) {
+            table.find('.year a').off('click').on('click', function (e) {
                 that._event = 'eventNext';
                 e.preventDefault();
                 e.stopPropagation();
@@ -3274,14 +3274,14 @@ $.widget("metro.calendar", {
                 that._mode = 'month';
                 that._renderCalendar();
             });
-            table.find('.btn-previous-year').on('click', function (e) {
+            table.find('.btn-previous-year').off('click').on('click', function (e) {
                 that._event = 'eventPrevious';
                 e.preventDefault();
                 e.stopPropagation();
                 that._distance -= 10;
                 that._renderCalendar();
             });
-            table.find('.btn-next-year').on('click', function (e) {
+            table.find('.btn-next-year').off('click').on('click', function (e) {
                 that._event = 'eventNext';
                 e.preventDefault();
                 e.stopPropagation();
@@ -3290,7 +3290,7 @@ $.widget("metro.calendar", {
             });
         }
 
-        table.find('.calendar-btn-today').on('click', function (e) {
+        table.find('.calendar-btn-today').off('click').on('click', function (e) {
             //that._event = 'eventNext';
             e.preventDefault();
             e.stopPropagation();
@@ -3301,7 +3301,7 @@ $.widget("metro.calendar", {
             that._day = that.options.date.getDate();
             that._renderCalendar();
         });
-        table.find('.calendar-btn-clear').on('click', function (e) {
+        table.find('.calendar-btn-clear').off('click').on('click', function (e) {
             e.preventDefault();
             e.stopPropagation();
             that.options.date = new Date();